Neela Pranavi/TalentLink-main/requirements-dev.txt
# Development dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.0.1
psycopg2-binary==2.9.7

# Development tools
black==23.9.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
coverage==7.3.2
